feat(README.md): add new banner for ATL Halloween 2025
Add a new banner image for ATL Halloween 2025 to the list of
available banners in the README. Update the total count of banners
from 26 to 27 to reflect the addition. This change ensures that
the documentation is up-to-date with the latest assets and provides
users with the most current visual resources.